% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_outputs.R
\name{B_outputs}
\alias{B_outputs}
\title{Our Meneu}
\usage{
B_outputs()
}
\value{
A dataframe with the implemented outputs
}
\description{
All the outputs possible. If you will use all outputs just type Allbasics( ) and set correctly all parameters. For multiple data frames list them and add g_ to any of the functions.
}
\examples{
#NOT RUN
# B_outputs()
}
