% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_arch_by_frame.R
\name{body_arch_by_frame}
\alias{body_arch_by_frame}
\title{Arch between three points}
\usage{
body_arch_by_frame(list_zones, sidesum = FALSE)
}
\arguments{
\item{list_zones}{A data frame or list from \link[CartesianBehaviour]{expdados} or \link[CartesianBehaviour]{list_zones}, which contains x and y for the three points of interest.}

\item{sidesum}{Define if the interest is the absolute angle for each side of the arc of the three points or the minimum and maximum angles. TRUE for minimum and maximum, FALSE by default.}
}
\value{
A data frame or lit with the body arch.
}
\description{
Arch between three points
}
\examples{
#Creating an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}

Traj <- TrajGenerate(10000, random = TRUE, stepLength = .01,
                     angularErrorSd = .8, fps = 30)

Traj <- cbind(Traj, data.frame(headx = rnorm(Traj$x), heady = rnorm(Traj$y),
              tailx = rnorm(Traj$x), taily = rnorm(Traj$y)))


#Create the data frame of the archs between three points
Angles <- body_arch_by_frame(Traj)
#View(Angles)
}
