% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ethovision}
\alias{Ethovision}
\title{Tracking positions from EthoVision for Leopard strain}
\format{
A list of 52 data frames
}
\source{
Laboratory zebrafish tracking experiment
}
\usage{
Ethovision
}
\description{
A named list of data frames, each with columns:
\describe{
\item{time}{Frame number or timestamp}
\item{x}{X position (pixels or mm)}
\item{y}{Y position (pixels or mm)}
}
}
\keyword{datasets}
