% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdados.R
\name{expdados}
\alias{expdados}
\title{Function to standardize a trajectory}
\usage{
expdados(
  dado,
  dadoz = NULL,
  Xaxi = "x_cm",
  Yaxi = "y_cm",
  Zaxi = NA,
  frames = "frame",
  id = 1,
  id_col = NULL,
  fps = 30
)
}
\arguments{
\item{dado}{The main data frame containing trajectory data (X and Y coordinates), an identifier of the subject and a frame or time information.}

\item{dadoz}{A second data frame that contains the z coordinates and the other identifiers.}

\item{Xaxi}{A character input representing the column in dado presenting the X coordinates.}

\item{Yaxi}{A character input representing the column in dado presenting the Y coordinates.}

\item{Zaxi}{Can be the name of the column in dadoz, or a vector of the type double (numeric) containing the Z coordinates.}

\item{frames}{The column in dados representing frame or time information. If this is the time column set the fps = 1.}

\item{id}{Numeric identifier of the subject.}

\item{id_col}{The column in dados that contain the subject identifier of the observations. If there is no identifier column all the data in the data frame will be considered.}

\item{fps}{The frame rate or another value to convert the time marker into a desired one.}
}
\value{
A standardize data frame. Note that if Z is a vector its length must be the same as of the X and Y.
}
\description{
Function to standardize a trajectory
}
\examples{
#Create an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}
Traj <- TrajGenerate(2000, random = TRUE, stepLength = .12,
                     angularErrorSd = .8, fps = 30)
#Export the trajectory
TrajEXP <- expdados(Traj, Xaxi = "x", Yaxi = "y",
                    frames = "time", id = 1, fps = 30)
#View(TrajEXP)

#Export the trajectory considering
#the Y axis both as Y and Z axis at the same time
TrajEXP2 <- expdados(Traj, Xaxi = "x", Yaxi = "y", Zaxi = "y",
                     frames = "time", id = 2, fps = 30)
#View(TrajEXP2)
}
