% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_list_zones.R
\name{g_list_zones}
\alias{g_list_zones}
\title{Split the trajectory in a list into zones of interest}
\usage{
g_list_zones(list_subjects, Zones, Zplane = NA, remove2D = F)
}
\arguments{
\item{list_subjects}{A list of data frames crated by \link[CartesianBehaviour]{g_expdados} or multiple \link[CartesianBehaviour]{expdados} listed containing the positions and time of an subject.}

\item{Zones}{A data frame crated by \link[CartesianBehaviour]{Zones_int} containing the zones of interest.}

\item{Zplane}{The complementary axis used in \link[CartesianBehaviour]{Zones_int}. Input can be "xz", "zx", "yz" or "zy".}

\item{remove2D}{To remove the 2D zones from the final data. FALSE by default.}
}
\value{
A list containing lists of splitted trajectory data frames.
}
\description{
This function is just a wrap around \link[CartesianBehaviour]{list_zones} with lapply. Check \link[CartesianBehaviour]{list_zones} for more info.
}
\examples{
#Create an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}

Traj <- list()
for(i in 1:10){
Traj[[i]] <- TrajGenerate(5000, random = TRUE, stepLength = .01,
                          angularErrorSd = .8, fps = 30)

}

#Export the trajectory.
TrajEXP <- g_expdados(Traj, Xaxi = "x", Yaxi = "y",
                      frames = "time", id = 1, fps = 30)

#Create the zones of interest
#Can be manually drawn but setting Zones = NA,
#but in this case is made by code directly
Z1 <- data.frame(x = c(-12,0,0,-12), y = c(10,10,0,0))
Z2 <- data.frame(x = c(0,12,12,0),   y = c(10,10,0,0))
Z3 <- data.frame(x = c(0,12,12,0),   y = c(0,0,-10,-10))
Z4 <- data.frame(x = c(-12,0,0,-12), y = c(0,0,-10,-10))
ZN <- list(Z1, Z2, Z3, Z4)
Zonas <- list(ZN)
rm(Z1, Z2, Z3, Z4, ZN)

#Split by the zones

TRAJ2D <- g_list_zones(TrajEXP, Zones = Zonas)
#View(TRAJ2D)
}
