% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_MMspeed.R
\name{g_MMspeed}
\alias{g_MMspeed}
\title{The maximum and minimum speeds for a list of subjects}
\usage{
g_MMspeed(list_group)
}
\arguments{
\item{list_group}{A list with the outputs of  \link[CartesianBehaviour]{speeds_by_frame} or \link[CartesianBehaviour]{Allbasics}.}
}
\value{
The maximum and minimum speeds for the zones of interest and total trajectory for multiple subject.
}
\description{
The maximum and minimum speeds for a list of subjects
}
\examples{
#Creating an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}

TRAJ <- list()
for(i in 1:10){
TRAJ[[i]] <- TrajGenerate(10000, random = TRUE, stepLength = .01,
                          angularErrorSd = .8, fps = 30)
}

#Create the zones of interest
#Can be manually drawn by setting Zones = NA,
#but in this case is made by code directly)
Z1 <- data.frame(x = c(-12,0,0,-12), y = c(10,10,0,0))
Z2 <- data.frame(x = c(0,12,12,0),   y = c(10,10,0,0))
Z3 <- data.frame(x = c(0,12,12,0),   y = c(0,0,-10,-10))
Z4 <- data.frame(x = c(-12,0,0,-12), y = c(0,0,-10,-10))
ZN <- list(Z1, Z2, Z3, Z4)
Zonas <- list(ZN)
rm(Z1, Z2, Z3, Z4, ZN, i)

#Perform all the measures
TRAJS2D <- g_Allbasics(TRAJ, Xaxi = "x",
                     Yaxi = "y", frames = "time", Zones = Zonas,
                     threshold = 3, Dist.reg = "point",
                     reg = data.frame(x = 0, y = 0))

 #Create the data frame of the
 #maximum and minimum speeds of multiple subjects
 Group.MaxMinSpeed <- g_MMspeed(TRAJS2D)
 #View(Group.MaxMinSpeed)
}
