% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_standardized_trj.R
\name{g_standardized_trj}
\alias{g_standardized_trj}
\title{Function to standardize a trajectory for a list}
\usage{
g_standardized_trj(
  list_subjects,
  dftz = NULL,
  Xaxi = "x_cm",
  Yaxi = "y_cm",
  Zaxi = NA,
  time = NA,
  frames = NA,
  id = 1,
  id_col = NULL,
  fps = 30
)
}
\arguments{
\item{list_subjects}{A list containing the main data frame containing trajectory data (X and Y coordinates), an identifier of the subject and a frame or time information.}

\item{dftz}{A second list that contains the data frames with the z coordinates and the other identifiers.}

\item{Xaxi}{A character input representing the column in list_subjects presenting the X coordinates.}

\item{Yaxi}{A character input representing the column in dft presenting the Y coordinates.}

\item{Zaxi}{Can be the name of the column in dftz, or a vector of the type double (numeric) containing the Z coordinates.}

\item{time}{The column in dfts representing time information.}

\item{frames}{The column in dfts representing frame information.}

\item{id}{Numeric identifier of the subject.}

\item{id_col}{The column in dfts that contain the subject identifier of the observations. If there is no identifier column all the data in the data frame will be considered.}

\item{fps}{The frame rate or another value to convert the time marker into a desired one.}
}
\value{
A standardized list of data frames
}
\description{
A standardize data frame. Note that if Z is a vector its length must be the same as of the X and Y.
}
\examples{
#Create an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}

Traj <- list()
for(i in 1:10){
Traj[[i]] <- TrajGenerate(5000, random = TRUE, stepLength = .01,
                          angularErrorSd = .8, fps = 30)

}

#Export the trajectory.
TrajEXP <- g_standardized_trj(Traj, Xaxi = "x", Yaxi = "y",
                      time = "time", id = 1, fps = 30)
#View(TrajEXP)
}
