% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency_zone.R
\name{latency_zone}
\alias{latency_zone}
\title{The latency to first enter a zone}
\usage{
latency_zone(list_zones)
}
\arguments{
\item{list_zones}{The output of \link[CartesianBehaviour]{list_zones} or \link[CartesianBehaviour]{Allbasics}.}
}
\value{
A data frame of the latency to first enter a zone of interest. if the subject starts in the zone the latency is 0. If it never reaches the zone the latency is the total time of observation.
}
\description{
The latency to first enter a zone
}
\examples{
#Creating an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}
Traj <- TrajGenerate(10000, random = TRUE, stepLength = .01,
                     angularErrorSd = .8, fps = 30)

#Create the zones of interest
#Can be manually drawn by setting Zones = NA,
#but in this case is made by code directly
Z1 <- data.frame(x = c(-12,0,0,-12), y = c(10,10,0,0))
Z2 <- data.frame(x = c(0,12,12,0),   y = c(10,10,0,0))
Z3 <- data.frame(x = c(0,12,12,0),   y = c(0,0,-10,-10))
Z4 <- data.frame(x = c(-12,0,0,-12), y = c(0,0,-10,-10))
ZN <- list(Z1, Z2, Z3, Z4)
Zonas <- list(ZN)
rm(Z1, Z2, Z3, Z4, ZN)

#Perform all the measures
TRAJ2D <- Allbasics(Traj, Xaxi = "x", Yaxi = "y", time = "time",
                    Zones = Zonas, threshold = 3,
                    Dist.reg = "point",
                    reg = data.frame(x = 0, y = 0))

#Create the data frame of the latency of the subject to first enter a zone
Latency.Zones <- latency_zone(TRAJ2D)
#View(Latency.Zones)
}
