% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_speed_zone.R
\name{mean_speed_zone}
\alias{mean_speed_zone}
\title{The mean speed for each zone}
\usage{
mean_speed_zone(dist_zone, time_zones, time_reescale = 1, dist_reescale = 1)
}
\arguments{
\item{dist_zone}{The output of \link[CartesianBehaviour]{dist_zone} or \link[CartesianBehaviour]{g_dist_zone}}

\item{time_zones}{The output of \link[CartesianBehaviour]{times_zone} or \link[CartesianBehaviour]{g_times_zone}}

\item{time_reescale}{Will divide the time to re-scale it.}

\item{dist_reescale}{Will divide the distances to re-scale it.}
}
\value{
A data frame with the mean speeds for total trajectory and for zones of interest.
}
\description{
The mean speed for each zone
}
\examples{
#Creating an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)
}
TRAJ <- list()

for(i in 1:10){
TRAJ[[i]] <- TrajGenerate(10000, random = TRUE, stepLength = .01,
                          angularErrorSd = .8, fps = 30)
}

#Create the zones of interest
#Can be manually drawn by setting Zones = NA,
#but in this case is made by code directly)
Z1 <- data.frame(x = c(-12,0,0,-12), y = c(10,10,0,0))
Z2 <- data.frame(x = c(0,12,12,0),   y = c(10,10,0,0))
Z3 <- data.frame(x = c(0,12,12,0),   y = c(0,0,-10,-10))
Z4 <- data.frame(x = c(-12,0,0,-12), y = c(0,0,-10,-10))
ZN <- list(Z1, Z2, Z3, Z4)
Zonas <- list(ZN)
rm(Z1, Z2, Z3, Z4, ZN)

#Perform all the mesures

  #For one subject
  TRJ <- Allbasics(dft = TRAJ, Xaxi = "x", Yaxi = "y", time = "time",
                    Zones = Zonas, threshold = 3, Dist.reg = "point",
                    reg = data.frame(x = 0, y = 0))

   #For multiple subjects

   #create an empty list to be filed with the trajectories

   TRAJS <- g_Allbasics(TRAJ, Xaxi = "x", Yaxi = "y", time = "time",
                         Zones = Zonas, threshold = 3,
                         Dist.reg = "point",
                         reg = data.frame(x = 0, y = 0))


rm(Zonas, TRAJ) #remove extra information

#Create the data frame of the mean speed in the zones for
#A group of subjects
Group.Time.Zones <- g_times_zone(TRAJS)
Group.Distance.Zones <- g_dist_zone(TRAJS)

G.Mean.Speed.Zones <- mean_speed_zone(Group.Distance.Zones,
                                     Group.Time.Zones)
#View(G.Mean.Speed.Zones)

#Create the data frame of the mean speed in the zones for one subject
Time.Zones <- times_zone(TRJ)
Distance.Zones <- dist_zone(TRJ)

Mean.Speed.Zones <- mean_speed_zone(Distance.Zones, Time.Zones)
#View(Mean.Speed.Zones)
}
