% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandering_zone.R
\name{meandering_zone}
\alias{meandering_zone}
\title{The meandering}
\usage{
meandering_zone(angle_zon, dist_zone, dist_reescale = 1)
}
\arguments{
\item{angle_zon}{The output of \link[CartesianBehaviour]{angle_zone} or \link[CartesianBehaviour]{g_angle_zone}.}

\item{dist_zone}{The output of \link[CartesianBehaviour]{dist_zone} or \link[CartesianBehaviour]{g_dist_zone}.}

\item{dist_reescale}{Will divide the distances to re-scale it.}
}
\value{
A data frame with the meandering of the subject(s).
}
\description{
The meandering
}
\examples{
#Creating an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)
}

TRAJ <- list()
for(i in 1:3){
TRAJ[[i]] <- TrajGenerate(10000, random = TRUE, stepLength = .01,
                          angularErrorSd = .8, fps = 30)
}

#Create the zones of interest
#Can be manually drawn by setting Zones = NA
#but in this case is made by code directly
Z1 <- base::data.frame(x = c(-12,0,0,-12), y = c(10,10,0,0))
Z2 <- base::data.frame(x = c(0,12,12,0),   y = c(10,10,0,0))
Z3 <- base::data.frame(x = c(0,12,12,0),   y = c(0,0,-10,-10))
Z4 <- base::data.frame(x = c(-12,0,0,-12), y = c(0,0,-10,-10))
ZN <- list(Z1, Z2, Z3, Z4)
Zonas <- list(ZN)
rm(Z1, Z2, Z3, Z4, ZN)

#Perform all the mesures

  #For one subject
  TRJ <- Allbasics(TRAJ[[1]], Xaxi = "x", Yaxi = "y", frames = "time",
                    Zones = Zonas, threshold = 3,
                    Dist.reg = "point",
                    reg = data.frame(x = 0, y = 0))

   #For multiple subjects
   TRAJS2D <- list() #create an empty list to be filed with the trajectories
   for (i in 1:length(TRAJ)) {

   TRAJS <- Allbasics(TRAJ[[i]], Xaxi = "x", Yaxi = "y", frames = "time",
                    Zones = Zonas, threshold = 3,
                    Dist.reg = "point",
                    reg = data.frame(x = 0, y = 0))
   Tj <- list(TRAJS)
   TRAJS2D <- append(TRAJS2D, Tj)

}
rm(Tj, Zonas, TRAJS, i, TRAJ) #remove extra information

#Create the data frame of the meandering
#inside of the zones for one subject
Angle.Zones <- angle_zone(TRJ)
Distance.Zones <- dist_zone(TRJ)

Meandering.Speed.Zones <- meandering_zone(Angle.Zones,
                                          Distance.Zones)
#View(Meandering.Speed.Zones)

#Create the data frame of the meandering
#of the zones for a group of subjects
Group.Angle.Zones <- g_angle_zone(TRAJS2D)
Group.Distance.Zones <- g_dist_zone(TRAJS2D)

Group.Meandering.Speed.Zones <- meandering_zone(Group.Angle.Zones,
                                                Group.Distance.Zones)
#View(Group.Meandering.Speed.Zones)
}
