% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToxTrac}
\alias{ToxTrac}
\title{Tracking positions from ToxTrac for Short‐fin strain}
\format{
A list of 47 data frames with columns \code{time}, \code{x}, \code{y}
}
\usage{
ToxTrac
}
\description{
Tracking positions from ToxTrac for Short‐fin strain
}
\keyword{datasets}
