% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajInPlanezone.R
\name{TrajInPlanezone}
\alias{TrajInPlanezone}
\title{Check if a Trajectory is inside or outside of a polygon considering the axis}
\usage{
TrajInPlanezone(Track, Zone, plane)
}
\arguments{
\item{Track}{A data frame containing points in two of x y or z axis of a trajectory.}

\item{Zone}{A data frame containing points in two of x y or z axis for a polygon.}

\item{plane}{the axis to be considered. Can be "zy", "yz", "zx", "xz", "xy" or "yx".}
}
\value{
A vector of 0's and 1's respectively for outside or inside the polygon.
}
\description{
This function is based on TrajInPolygon function from trajr and which apply the function point.in.polygon from sp.
}
\examples{
Traj <- matrix(rnorm(2400,0,4), ncol=3,
dimnames  = list(c(1:800), c("x", "y", "z")))
Traj <- data.frame(Traj)
Polxy <- data.frame(x = c(5, 5, -5, -5), y = c(4, -4, -4, 4))
Polyz <- data.frame(y = c(4, 4, -4, -4), z = c(7, -7, -7, 7))
Polxz <- data.frame(x = c(5, 5, -5, -5), z = c(7, -7, -7, 7))

Teste1 <- data.frame(TrajInPlanezone(Traj, Polxy, "yx"))
#View(Teste1)

Teste2 <- data.frame(TrajInPlanezone(Traj, Zone = Polyz, plane = "yz"))
#View(Teste2)

Teste3 <- data.frame(TrajInPlanezone(Traj, Zone = Polxz, plane = "xz"))
#View(Teste3)
}
