% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_sepisods.R
\name{under_sepisods}
\alias{under_sepisods}
\title{Episodes & time under a defined speed and above a time period}
\usage{
under_sepisods(list_zones, time_threshold = 1, MOD = "USER")
}
\arguments{
\item{list_zones}{The output of \link[CartesianBehaviour]{threshold_speed_by_frame} or \link[CartesianBehaviour]{Allbasics}}

\item{time_threshold}{The minimum time for an episode to be counted.}

\item{MOD}{Don't change, for internal use in \link[CartesianBehaviour]{over_time_ana}.}
}
\value{
The time the episodes lasted and the number of episodes completely in the zones and for total track. Note that even if the zones are tangents the sum of that can be greater or lesser than the number of episodes in the total trajectory.
}
\description{
Episodes & time under a defined speed and above a time period
}
\examples{
#Creating an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)
}
Traj <- TrajGenerate(10000, random = TRUE, stepLength = .01,
                      angularErrorSd = .8, fps = 30)

#Create the zones of interest
#Can be manually drawn by setting Zones = NA,
#but in this case is made by code directly
Z1 <- data.frame(x = c(-12,0,0,-12), y = c(10,10,0,0))
Z2 <- data.frame(x = c(0,12,12,0),   y = c(10,10,0,0))
Z3 <- data.frame(x = c(0,12,12,0),   y = c(0,0,-10,-10))
Z4 <- data.frame(x = c(-12,0,0,-12), y = c(0,0,-10,-10))
ZN <- list(Z1, Z2, Z3, Z4)
Zonas <- list(ZN)
rm(Z1, Z2, Z3, Z4, ZN)

#Perform all the measures
TRAJ2D <- Allbasics(Traj, Xaxi = "x", Yaxi = "y", time = "time",
                    Zones = Zonas, threshold = 3,
                    Dist.reg = "point",
                    reg = data.frame(x = 0, y = 0))

#Create the data frame of the episodes under a
#defined speed of the subject
Under.Speed.Episods <- under_sepisods(TRAJ2D, time_threshold = 0.001)
#View(Under.Speed.Episods)
}
