% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zones_int.R
\name{Zones_int}
\alias{Zones_int}
\title{Create zones of interest}
\usage{
Zones_int(
  nzonesd2,
  nzonesd3 = 0,
  faceZ,
  maxX,
  minX,
  maxY,
  minY,
  maxZ,
  minZ,
  npts = 4,
  Circ_Arena = FALSE
)
}
\arguments{
\item{nzonesd2}{The number of zones, for 2D or the XY dimension for 3D analysis.}

\item{nzonesd3}{The number of zones To be drawn in the Z and complementary axis. Note that the 3D zones will be a  combination of the X-Y and Z-complementary dimensions. Also is impotent to note that the 2D zones will be generated and analyzed.}

\item{faceZ}{The complementary axis which will be used to draw the zones in the 3D analysis. Input 1 for x, 2 for y.}

\item{maxX}{The maximum  value on the X-axis in which movement was recorded.}

\item{minX}{The minimum value on the X-axis in which the movement was recorded.}

\item{maxY}{The maximum value on the Y-axis in which movement was recorded.}

\item{minY}{The minimum value on the Y-axis in which the movement was recorded.}

\item{maxZ}{The maximum value on the Z-axis in which movement was recorded.}

\item{minZ}{The minimum value on the Z-axis in which the movement was recorded.}

\item{npts}{Number of vertices that define the polygon representing the zone of interest.}

\item{Circ_Arena}{If TRUE draws a circle to represent an arena where the movement was recorded. FALSE by default.}
}
\value{
A list containing the 1 (for 2d) or 2 (for 3d) data frames containing the vertices of the zones of interest. The polygon is made by pointing in the plot the points, then a pop-up allows editing the inputs of points.
}
\description{
Create zones of interest
}
\examples{
#Create 3 2D zones of interest in a arena with measuring 20x10
#and Cartesian origin in the center of the arena
#zons2D <- Zones_int(nzonesd2 = 3, maxX = 10, minX = -10, maxY = 5,
#                    minY = -5, npts = 4, Circ_Arena = TRUE)
#Create 2 zones of interest in the XY axis and 4 zones in the ZX axis in
#an arena with measuring 20x10x14 and Cartesian origin in the center of the arena
#zons3D <- Zones_int(nzonesd2 = 2, nzonesd3 = 4, faceZ = 1, maxX = 10,
#                    minX = -10, maxY = 5, minY = -5, maxZ = 7, minZ = -7, npts = 4)
}
