% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdados.R
\name{expdados}
\alias{expdados}
\title{Function to standardize a trajectory}
\usage{
expdados(
  dado,
  dadoz = NULL,
  Xaxi = "x_cm",
  Yaxi = "y_cm",
  Zaxi = NA,
  frames = NA,
  time = NA,
  id = 1,
  id_col = NULL,
  fps = 30,
  Gtype = "unique",
  center = NA,
  head = NA,
  tail = NA
)
}
\arguments{
\item{dado}{The main data frame containing trajectory data (X and Y coordinates), an identifier of the subject and a frame or time information.}

\item{dadoz}{A second data frame that contains the z coordinates and the other identifiers. The first frame of data must be the same as the dado input.}

\item{Xaxi}{A character input representing the column in dado presenting the X coordinates.}

\item{Yaxi}{A character input representing the column in dado presenting the Y coordinates.}

\item{Zaxi}{Can be the name of the column in dadoz, or a vector of the type double (numeric) containing the Z coordinates.}

\item{frames}{The column in dados representing frame information.}

\item{time}{The column in dados representing time information.}

\item{id}{Numeric identifier of the subject.}

\item{id_col}{The column in dados that contain the subject identifier of the observations. If there is no identifier column all the data in the data frame will be considered.}

\item{fps}{The frame rate or another value to convert the time marker into a desired one.}

\item{Gtype}{If you have any kind of of 3 point tracking select a type of data input. "unique" by default. See more in description.}

\item{center}{The name of the center point input.}

\item{head}{The name of the head columns or a data frame containing it.}

\item{tail}{The name of the tail columns or a data frame containing it.}
}
\value{
A standardize data frame. Note that if Z is a vector its length must be the same as of the X and Y.
}
\description{
To standardize data from 1 or multiple data frames. For the 3 point positions of interest to future mesure bodi arch angles you have to chose between
"1Lcol3Point" - which is a dataframe with 1 column names and is just straight forward name  the column identifier in the head, center, tail parameters.
"2Lcol3Point" - optimized for DeepLabCut outputs, which have 3 lines of columns, 1 with a generated unique name, and 2 lines for defining the body parts and the type of coordinates (X, Y, Z), separately. in this case it will paste the "X/Y" + _ + "Body part column name" to perform the anlisis.
"multiDF3Point" - If you have the body parts spread in 3 different data frames. note that the center must be in the dados input.
}
\examples{
#Create an trajectory using the function from the package trajr
if("trajr" \%in\% rownames(installed.packages()) == FALSE){
install.packages("trajr")}else{library(trajr)}
Traj <- TrajGenerate(2000, random = TRUE, stepLength = .12,
                     angularErrorSd = .8, fps = 30)
#Export the trajectory
TrajEXP <- expdados(Traj, Xaxi = "x", Yaxi = "y",
                    time = "time", id = 1, fps = 30)
#View(TrajEXP)

#Export the trajectory considering
#the Y axis both as Y and Z axis at the same time
TrajEXP2 <- expdados(Traj, Xaxi = "x", Yaxi = "y", Zaxi = "y",
                     time = "time", id = 2, fps = 30)
#View(TrajEXP2)
}
